restAPI/services/calendar/requirements.txt
# Some dependencies are already install by restAPI-package

python-dotenv==1.0.0
